--- playwright-bdd/playwright.config.js
@@ -1,38 +1,0 @@
-import { defineConfig, devices } from '@playwright/test';
-import { defineBddConfig } from 'playwright-bdd';
-
-const testDir = defineBddConfig({
-  paths: ['features/*.feature'],
-  require: ['steps/*.js'],
-});
-
-export default defineConfig({
-  testDir,
-  /* timeout */
-  timeout: 30 * 1000,
-  /* Run tests in files in parallel */
-  fullyParallel: true,
-  /* Reporter to use. See https://playwright.dev/docs/test-reporters */
-  reporter: 'html',
-  use: {
-    /* Base URL to use in actions like `await page.goto('/')`. */
-    // baseURL: 'http://127.0.0.1:3000',
-
-    /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
-    trace: 'on-first-retry',
-    baseURL: 'http://localhost:61100/'
-  },
-
-  /* Configure projects for major browsers */
-  projects: [
-    //{ name: 'setup', testMatch: /.*\.setup\.js/ },
-    {
-      name: 'chromium',
-      use: { ...devices['Desktop Chrome'],
-      // Use prepared auth state.
-      storageState: 'auth.json',
-     },
-     //dependencies: ['setup'],
-    },
-  ],
-});
